text_rpg/requirements.txt
streamlit>=1.32.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
PyMySQL>=1.1.0
bcrypt>=4.0.0
